python/requirements-deploy.txt
protobuf>=4,<5
grpcio==1.62.2
grpcio-tools==1.62.2
docker==7.1.0
pyyaml==6.0.1
kubernetes==30.1.0